Run the rollup backfill scan from the worker instead of the scheduler
The scheduler app has no S3 credentials, so listing archived days from
it failed every tick with ECONNREFUSED against the localhost fallback
and no backfill job was ever enqueued. The scheduler now enqueues a
single parameterless scan job, and the worker lists the archive, picks
the missing days, and backfills them oldest first.
